projects/heart-disease-risk/fn-ai-ml-heart-disease/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
scikit-learn==1.2.2
